+++ processor/src/test/fixtures/expected/com/example/post_construct/Arez_ActionPostConstructModel.java
@@ -1,0 +1,125 @@
+package com.example.post_construct;
+
+import arez.ActionFlags;
+import arez.Arez;
+import arez.ArezContext;
+import arez.Component;
+import arez.ComputableValue;
+import arez.Disposable;
+import arez.SafeProcedure;
+import arez.component.DisposeNotifier;
+import arez.component.Identifiable;
+import arez.component.internal.ComponentKernel;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import org.realityforge.braincheck.Guards;
+
+@Generated("arez.processor.ArezProcessor")
+@SuppressWarnings("unchecked")
+final class Arez_ActionPostConstructModel extends ActionPostConstructModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
+  private static volatile int $$arezi$$_nextId;
+
+  private final ComponentKernel $$arezi$$_kernel;
+
+  @Nonnull
+  private final ComputableValue<Integer> $$arez$$_someValue;
+
+  Arez_ActionPostConstructModel() {
+    super();
+    final ArezContext $$arezv$$_context = Arez.context();
+    final int $$arezv$$_id = ( Arez.areNamesEnabled() || Arez.areRegistriesEnabled() || Arez.areNativeComponentsEnabled() ) ? ++$$arezi$$_nextId : 0;
+    final String $$arezv$$_name = Arez.areNamesEnabled() ? "ActionPostConstructModel." + $$arezv$$_id : null;
+    final Component $$arezv$$_component = Arez.areNativeComponentsEnabled() ? $$arezv$$_context.component( "ActionPostConstructModel", $$arezv$$_id, $$arezv$$_name, () -> $$arezi$$_nativeComponentPreDispose() ) : null;
+    this.$$arezi$$_kernel = new ComponentKernel( Arez.areZonesEnabled() ? $$arezv$$_context : null, Arez.areNamesEnabled() ? $$arezv$$_name : null, $$arezv$$_id, Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, null, Arez.areNativeComponentsEnabled() ? null : this::$$arezi$$_dispose, null, true, false, false );
+    this.$$arez$$_someValue = $$arezv$$_context.computable( Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, Arez.areNamesEnabled() ? $$arezv$$_name + ".someValue" : null, () -> super.someValue(), ComputableValue.Flags.AREZ_DEPENDENCIES | ComputableValue.Flags.RUN_LATER );
+    this.$$arezi$$_kernel.componentConstructed();
+    this.postConstruct();
+    this.$$arezi$$_kernel.componentReady();
+  }
+
+  private int $$arezi$$_id() {
+    return this.$$arezi$$_kernel.getId();
+  }
+
+  @Override
+  @Nonnull
+  public Integer getArezId() {
+    return $$arezi$$_id();
+  }
+
+  private void $$arezi$$_nativeComponentPreDispose() {
+    this.$$arezi$$_kernel.notifyOnDisposeListeners();
+  }
+
+  @Override
+  public void addOnDisposeListener(@Nonnull final Object key, @Nonnull final SafeProcedure action) {
+    this.$$arezi$$_kernel.addOnDisposeListener( key, action );
+  }
+
+  @Override
+  public void removeOnDisposeListener(@Nonnull final Object key) {
+    this.$$arezi$$_kernel.removeOnDisposeListener( key );
+  }
+
+  @Override
+  public boolean isDisposed() {
+    return this.$$arezi$$_kernel.isDisposed();
+  }
+
+  @Override
+  public void dispose() {
+    this.$$arezi$$_kernel.dispose();
+  }
+
+  private void $$arezi$$_dispose() {
+    this.$$arez$$_someValue.dispose();
+  }
+
+  @Override
+  void postConstruct() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'postConstruct' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + ( null == this.$$arezi$$_kernel ? "?" : this.$$arezi$$_kernel.getName() ) + "'" );
+    }
+    this.$$arezi$$_kernel.getContext().safeAction(Arez.areNamesEnabled() ? this.$$arezi$$_kernel.getName() + ".postConstruct" : null, () -> super.postConstruct(), ActionFlags.READ_WRITE | ActionFlags.VERIFY_ACTION_REQUIRED, null );
+  }
+
+  @Override
+  public int someValue() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'someValue' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + ( null == this.$$arezi$$_kernel ? "?" : this.$$arezi$$_kernel.getName() ) + "'" );
+    }
+    return this.$$arez$$_someValue.get();
+  }
+
+  @Override
+  public int hashCode() {
+    if ( Arez.areNativeComponentsEnabled() ) {
+      return Integer.hashCode( $$arezi$$_id() );
+    } else {
+      return super.hashCode();
+    }
+  }
+
+  @Override
+  public boolean equals(final Object o) {
+    if ( Arez.areNativeComponentsEnabled() ) {
+      if ( o instanceof Arez_ActionPostConstructModel ) {
+        final Arez_ActionPostConstructModel that = (Arez_ActionPostConstructModel) o;
+        return $$arezi$$_id() == that.$$arezi$$_id();
+      } else {
+        return false;
+      }
+    } else {
+      return super.equals( o );
+    }
+  }
+
+  @Override
+  public String toString() {
+    if ( Arez.areNamesEnabled() ) {
+      return "ArezComponent[" + this.$$arezi$$_kernel.getName() + "]";
+    } else {
+      return super.toString();
+    }
+  }
+}