+++ processor/src/test/fixtures/input/com/example/post_construct/ActionPostConstructModel.java
@@ -1,0 +1,22 @@
+package com.example.post_construct;
+
+import arez.annotations.Action;
+import arez.annotations.ArezComponent;
+import arez.annotations.Memoize;
+import arez.annotations.PostConstruct;
+
+@ArezComponent
+abstract class ActionPostConstructModel
+{
+  @Action
+  @PostConstruct
+  void postConstruct()
+  {
+  }
+
+  @Memoize
+  public int someValue()
+  {
+    return 0;
+  }
+}